fix missing instance panic

+++ internal/controller/intelmachine_controller.go
@@ -224,15 +224,19 @@ func (r *IntelMachineReconciler) reconcileDelete(rc IntelMachineReconcilerContex
 
 	if controllerutil.ContainsFinalizer(rc.intelMachine, infrastructurev1alpha1.FreeInstanceFinalizer) {
 		// Remove the instance from the workload in Inventory
-		req := inventory.DeleteInstanceFromWorkloadInput{
-			TenantId:   rc.intelCluster.Namespace,
-			WorkloadId: rc.intelCluster.Spec.ProviderId,
-			InstanceId: *rc.intelMachine.Spec.ProviderID,
-		}
-		res := r.InventoryClient.DeleteInstanceFromWorkload(req)
-		if res.Err != nil && !errors.Is(res.Err, inventory.ErrInvalidWorkloadMembers) {
-			rc.log.Error(res.Err, "Failed to delete instance from workload in Inventory")
-			return res.Err
+		if rc.intelMachine.Spec.ProviderID != nil {
+			req := inventory.DeleteInstanceFromWorkloadInput{
+				TenantId:   rc.intelCluster.Namespace,
+				WorkloadId: rc.intelCluster.Spec.ProviderId,
+				InstanceId: *rc.intelMachine.Spec.ProviderID,
+			}
+			res := r.InventoryClient.DeleteInstanceFromWorkload(req)
+			if res.Err != nil && !errors.Is(res.Err, inventory.ErrInvalidWorkloadMembers) {
+				rc.log.Error(res.Err, "Failed to delete instance from workload in Inventory")
+				return res.Err
+			}
+		} else {
+			rc.log.Info("ProviderID is nil, skipping instance deletion from workload")
 		}
 		controllerutil.RemoveFinalizer(rc.intelMachine, infrastructurev1alpha1.FreeInstanceFinalizer)
 	}
